+++ src/main/java/com/example/mytunes/HelloApplication.java
@@ -12,7 +12,7 @@ public class HelloApplication extends Application {
     public void start(Stage stage) throws IOException {
         FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("hello-view.fxml"));
         Scene scene = new Scene(fxmlLoader.load(), 972, 665);
-        stage.setTitle("wtaheva!");
+        stage.setTitle("Hello!");
         stage.setScene(scene);
         stage.show();
     }
