--- src/main/java/com/example/mytunes/HelloController.java
@@ -1,27 +1,0 @@
-package com.example.mytunes;
-
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-
-public class HelloController {
-    @FXML
-    private Label welcomeText;
-
-
-    @FXML
-    public void playlistNewButtonAction(ActionEvent event) {
-    }
-    @FXML
-    public void playlistEditButtonAction(ActionEvent event) {
-    }
-    @FXML
-    public void deleteButtonAction(ActionEvent event) {
-    }
-    @FXML
-    public void songEditButtonAction(ActionEvent event) {
-    }
-    @FXML
-    public void songNewButtonAction(ActionEvent event) {
-    }
-}
