--- src/main/java/com/example/mytunes/songController.java
@@ -1,33 +1,0 @@
-package com.example.mytunes;
-
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-
-public class songController {
-    @FXML
-    public Label songName;
-    @FXML
-    public Label artistName;
-    @FXML
-    public void addNewSong(ActionEvent actionEvent) {
-    }
-    @FXML
-    public void playCurrentSong(ActionEvent actionEvent) {
-    }
-    @FXML
-    public void stop(ActionEvent actionEvent) {
-    }
-    @FXML
-    public void playPreviousSong(ActionEvent actionEvent) {
-    }
-    @FXML
-    public void playNextSong(ActionEvent actionEvent) {
-    }
-    @FXML
-    public void addNewPlaylist(ActionEvent actionEvent) {
-    }
-    @FXML
-    public void switchToAllSongs(ActionEvent actionEvent) {
-    }
-}
