+++ src/main/java/module-info.java
@@ -5,4 +5,6 @@ module com.example.mytunes {
 
     opens com.example.mytunes to javafx.fxml;
     exports com.example.mytunes;
+    exports com.example.mytunes.gui.controller;
+    opens com.example.mytunes.gui.controller to javafx.fxml;
 }
