+++ src/main/resources/com/example/mytunes/hello-view.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.layout.*?>
 <?import javafx.scene.text.*?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="665.0" prefWidth="972.0" style="-fx-background-color: #F4A460; -fx-background-radius: 5;" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.mytunes.HelloController">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="665.0" prefWidth="972.0" style="-fx-background-color: #F4A460; -fx-background-radius: 5;" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.mytunes.gui.controller.HelloController">
    <children>
       <TableView fx:id="allSongsTableView" layoutX="600.0" layoutY="56.0" prefHeight="269.0" prefWidth="344.0" style="-fx-background-radius: 5; -fx-background-color: #F0F0F0;">
         <columns>
