--- src/main/resources/com/example/mytunes/songView.fxml
@@ -1,98 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ListView?>
-<?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-
-<BorderPane prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.mytunes.songController">
-   <center>
-      <ListView prefHeight="200.0" prefWidth="200.0" BorderPane.alignment="CENTER">
-         <BorderPane.margin>
-            <Insets bottom="5.0" right="5.0" top="5.0" />
-         </BorderPane.margin></ListView>
-   </center>
-   <top>
-      <HBox prefHeight="60.0" prefWidth="800.0" BorderPane.alignment="CENTER">
-         <children>
-            <VBox alignment="CENTER_LEFT" prefHeight="100.0" prefWidth="184.0">
-               <children>
-         
-                       <Label fx:id="songName" text="song">
-                     <font>
-                        <Font name="System Bold" size="14.0" />
-                     </font>
-                  </Label>
-                  <Label fx:id="artistName" layoutX="10.0" layoutY="52.0" text="artist" />
-               </children>
-               <HBox.margin>
-                  <Insets left="5.0" />
-               </HBox.margin>
-            </VBox>
-            <VBox layoutX="410.0" layoutY="10.0" prefHeight="200.0" prefWidth="100.0" />
-            <VBox alignment="CENTER" layoutX="333.0" layoutY="10.0" prefHeight="59.0" prefWidth="70.0">
-               <children>
-                  <Button mnemonicParsing="false" onAction="#playPreviousSong" text="Previous" />
-               </children>
-            </VBox>
-            <VBox alignment="CENTER" layoutX="155.0" layoutY="10.0" prefHeight="59.0" prefWidth="70.0">
-               <children>
-                  <Button alignment="TOP_LEFT" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" mnemonicParsing="false" onAction="#playCurrentSong" text="Play" />
-               </children>
-            </VBox>
-            <VBox alignment="CENTER" layoutX="455.0" layoutY="10.0" prefHeight="59.0" prefWidth="70.0">
-               <children>
-                  <Button mnemonicParsing="false" onAction="#stop" text="Stop" />
-               </children>
-            </VBox>
-            <VBox alignment="CENTER" layoutX="555.0" layoutY="10.0" prefHeight="59.0" prefWidth="70.0">
-               <children>
-                  <Button mnemonicParsing="false" onAction="#playNextSong" text="Next" />
-               </children>
-            </VBox>
-            <VBox alignment="CENTER" layoutX="610.0" layoutY="10.0" prefHeight="200.0" prefWidth="100.0" />
-            <VBox alignment="CENTER" layoutX="480.0" layoutY="10.0" prefHeight="200.0" prefWidth="100.0">
-               <children>
-                 <Button onAction="#addNewSong" text="New Song" />
-               </children>
-            </VBox>
-         </children>
-      </HBox>
-   </top>
-   <left>
-      <VBox alignment="CENTER" prefHeight="474.0" prefWidth="100.0" BorderPane.alignment="CENTER">
-         <children>
-            <HBox alignment="CENTER" prefHeight="45.0" prefWidth="100.0">
-               <children>
-                  <Button mnemonicParsing="false" onAction="#switchToAllSongs" text="All Songs" />
-               </children>
-            </HBox>
-            <HBox alignment="CENTER" prefHeight="45.0" prefWidth="100.0">
-               <children>
-                  <Label text="Playlists" />
-               </children>
-            </HBox>
-            <ListView prefHeight="400.0" prefWidth="90.0">
-               <VBox.margin>
-                  <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-               </VBox.margin>
-            </ListView>
-            <Button alignment="BOTTOM_CENTER" mnemonicParsing="false" onAction="#addNewPlaylist" text="New Playlist">
-               <VBox.margin>
-                  <Insets top="1.0" />
-               </VBox.margin>
-            </Button>
-         </children>
-         <BorderPane.margin>
-            <Insets top="5.0" />
-         </BorderPane.margin>
-         <padding>
-            <Insets top="-5.0" />
-         </padding>
-      </VBox>
-   </left>
-</BorderPane>
